joy2key: allow using legacy TIOCSTI method for emitting keyboard events
On desktop platforms, `uinput` sends keyboard events to the window in the foreground. If that window isn't the terminal where `runcommand` is running, its menu cannot be triggered. This happens when, for example, a launch image is being displayed (the image viewer will be the the one getting the keyboard events).

On Linux distributions that disable TIOCSTI (e.g. Ubuntu 24.04), using `uinput` is the only option. But if TIOCSTI is available, it can be used instead by setting `JOY2KEY_USE_LEGACY_TIOCSTI=1` or invoking with the `--legacy-tiocsti`/`-t` parameter.