mupen64plus - de-duplicate make parameters
The make parameters are needed at the install stage and build stage so move them to a function.

Rename variable $source to $dir in the install loop - This matches the build function. An undefined variable $dir was used previously.

This also fixes some slight differences between them that were unnecessary:
  * isPlatform "rpi1" && params+=("VFP=1" "VFP_HARD=1") was missing from the install stage
  * install was missing isPlatform "armv8" && params+=("HOST_CPU=armv8") which is needed when building for the rpi5 with 64bit kernel and 32bit userland.