issues: new issue template
Leverage the new Github features for issues by adding an extra stop to point users to forums/docs when attempting to open an issue.
The new issue template was modified to include the previous note, but now formatted with Markdown.

Removed the ISSUE_TEMPLATE.md since it's been deprecated by Github since March 2025:
<https://github.blog/changelog/2025-02-18-github-issues-projects-february-18th-update/#%f0%9f%8c%85-single-issue-templates-issue_template-md-will-be-retired>