joy2key: match config by Product/Vendor IDs
When trying to match a RetroArch configuration profile to a detected joystick, try to use the Product/Vendor IDs in addition to Name and GUID. This covers the cases where the device naming changes between profile creation and `joy2key_sdl` runtime. We previously matched by the `input_device_name` found in the `.cfg` and GUID, but the GUID was actually not useful since it can't be read from the config profile `.cfg`.

Added also a few more debugging statements to show the configuration profiles loaded and to print the Product/Vendor IDs along the device name.