dkms: amend removal of module for newer DKMS
The `status` command returns `module/ver` in the first column, so amend the parsing to correctly extract the module version. Since the listing in newer version is a mix of `/` and `,` as separators, use `awk` to parse the line.

Should fix the uninstallation of drivers in recent latest RaspiOS/Debian (Bookworm, 12).