various: don't depend on 'libraspberrypi'
Remove the dependency on `libraspberrypi` packages for new RaspiOS Bookworm installations, since they're not useful anymore.

To do this, make the installation/usage in various scriptmodules depend on the `dispmanx` system flag instead of the `rpi` system flag. `dispmanx` is not set/used on RaspiOS Bookworm/Bullseye, but will still be set/used on Buster as before for Pi 0-4 platforms.