raspbiantools: replace sdl1 package on upgrades when not needed
Our custom SDL1 package is not needed on non-Dipmanx platforms, but due to the previously installation flag ('rpi') could still be installed and depend on `libraspberrypi` (which is deprecated). So on `rpi` systems where `dispmanx` is not available and thus our package is not needed, replace the installed package with the version from the upstream repository. This should happen when an update + system packages is used.