various: don't depend on 'libraspberrypi' (#4085)
* various: don't depend on 'libraspberrypi'

Remove the dependency on `libraspberrypi` packages for new RaspiOS Bookworm installations, since they're not useful anymore.

To do this, make the installation/usage in various scriptmodules depend on the `dispmanx` system flag instead of the `rpi` system flag. `dispmanx` is not set/used on RaspiOS Bookworm/Bullseye, but will still be set/used on Buster as before for Pi 0-4 platforms.

* raspbiantools: replace sdl1 package on upgrades when not needed

Our custom SDL1 package is not needed on non-Dipmanx platforms, but due to the previously installation flag ('rpi') could still be installed and depend on `libraspberrypi` (which is deprecated). So on `rpi` systems where `dispmanx` is not available and thus our package is not needed, replace the installed package with the version from the upstream repository. This should happen when an update + system packages is used.