minivmac: change the upstream source and build
The original project at https://www.gryphel.com/c/minivmac/ is no longer active and the community has created https://github.com/minivmac to continue the development of the emulator.
This updated version has SDL2 support and is still updated.

Use the new upstream project as the source for our package instead of the SDL2 fork we were using (which is no longer updated) and adjust the build process for it.