love: enable for 64bit ARM
Building for ARM64 was disabled due to a build failure on Odroid back in 2016 (commit-ish: acba46e0953ac7). But since it builds fine on modern 64bit Debian/Ubuntu (including RaspiOS), enable it again.

Tested build on Trixie, games seems to run fine (_mari0_ and the game posted in the forum/topic/37242/).